requirements.txt
google-cloud-error-reporting==0.30.1
pillow==5.4.1
numpy==1.16.2
google-cloud-storage